pdf_processing/requirements.txt
pymupdf==1.25.2
psycopg2-binary==2.9.9
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pytest==7.4.4
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6